% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandleau.R
\docType{data}
\name{pandleau}
\alias{pandleau}
\title{global reference to pandleau (will be initialized in .onLoad)}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
pandleau
}
\description{
global reference to pandleau (will be initialized in .onLoad)
}
\keyword{datasets}
