% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{publish_tde}
\alias{publish_tde}
\title{Publish a Tableau Data Extract (.tde) file to a Tableau Server
 using the Tableau SDK.}
\usage{
publish_tde(filename, datasource, hostname, username, password, siteID, project,
  overwrite = FALSE)
}
\arguments{
\item{filename}{a valid .tde file to publish}

\item{datasource}{what you want to name your extract on the server}

\item{hostname}{the full hostname of your Tableau Server}

\item{username}{Tableau Server username}

\item{password}{Tableau Server password}

\item{siteID}{The site ID, use empty string for the default site}

\item{project}{The name of the project on Tableau Server to publish to.
If unspecified it will publish to the default project}

\item{overwrite}{logical. If TRUE, any existing extract with the
same datasource name is destroyed. If FALSE, it will only publish
if the datasource name does not already exist.}
}
\description{
Publish a Tableau Data Extract (.tde) file to a Tableau Server
 using the Tableau SDK.
}
\examples{
publish_tde(iris, "iris.tde", "Iris", "http://localhost", "username", "password")

}

