% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tableau.R
\name{write_tableau}
\alias{write_tableau}
\title{Write R data.frame to Tableau Extract.}
\usage{
write_tableau(data, path, spatial_columns = NA, spatial_indicator = T,
  add_index = F)
}
\arguments{
\item{data}{input data.frame}

\item{path}{output path of Tableau Extract}

\item{spatial_columns}{index of spatial column, either number or name}

\item{spatial_indicator}{change spatial characteristic}

\item{add_index}{Add index column to Tableau Extract}
}
\description{
Write R data.frame to Tableau Extract.
}
\examples{
write_tableau(iris,"iris.tde")
}
