% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tde.R
\name{write_tde}
\alias{write_tde}
\title{Write a R data.frame as a Tableau Data Extract (.tde) file using
the Tableau SDK.}
\usage{
write_tde(df, filename, append = FALSE)
}
\arguments{
\item{df}{the data.frame to convert}

\item{filename}{the output of it (must contain .tde)}

\item{append}{logical. If TRUE, append data to the file.
If FALSE, any existing file of the name is destroyed.}
}
\description{
Write a R data.frame as a Tableau Data Extract (.tde) file using
the Tableau SDK.
}
\examples{
write_tde(iris, "iris.tde")

}

