% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tableau.R
\name{write_tableau}
\alias{write_tableau}
\title{Write R data.frame to Tableau Extract.}
\usage{
write_tableau(data, path, spatial_columns = NA, spatial_indicator = T,
  add_index = F)
}
\arguments{
\item{data}{input data.frame}

\item{path}{output path of Tableau Extract}

\item{spatial_columns}{index of spatial column, either number or name}

\item{spatial_indicator}{change spatial characteristic}

\item{add_index}{Add index column to Tableau Extract}
}
\description{
Write R data.frame to Tableau Extract.
}
\examples{
\dontrun{
# The most simple example is outputting iris dataset into .tde or .hyper format:
write_tableau(iris,"iris.tde")


# Another example that involves setting spatial column and add index:
# I grabbed the following Brazil flights data off of kaggle for this example:
https://www.kaggle.com/microtang/exploring-brazil-flights-data/data.
library(tidyverse)
library(RTableau)
# Read in the data and select first 1000 rows
example_df <- read_csv("BrFlights2.csv")
example_df <- slice(example_df,1:1000)

# Set up a spatial column as `POINT (Lon Lat)` format
example_df <- example_df \%>\%
  mutate(SpatialDest = paste0("POINT (",round(LongDest,6)," ",round(LatDest,6),")"))
# Write .tde or .hyper Extract!
write_tableau(example_df,"BrFlights.tde","SpatialDest",add_index = T)
}
}
